Improve make_bit example
Use the '-z' option of the 'test' command to check whether 'MSG' is
a string of zero length.

Do not test the variable 'MSG' twice, return '1' by using the
'false' command in the first 'if' statement. In this way the 'push'
target is aborted right after printing the reminder. There is no
need to test it once again.